% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_row_numbers.R
\name{add_row_numbers}
\alias{add_row_numbers}
\title{Agregar numero de fila}
\usage{
add_row_numbers(dat, .rows_name = NULL)
}
\arguments{
\item{dat}{Un data.frame}

\item{.rows_name}{(chr) Nombre para la columna}
}
\value{
Un data.frame
}
\description{
Agrega una primera columna que contiene los numeros de fila y el nombre
especificado. Pensado como paso final para tablas a renderizar en R Markdown.
}
\examples{
add_row_numbers(iris, "n_fila")
}
