% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_api.R
\name{gh_repos_issues}
\alias{gh_repos_issues}
\title{Obtener issues de un repositorio de Github}
\usage{
gh_repos_issues(owner = "calderonsamuel", repo = ".", state = "open", ...)
}
\arguments{
\item{owner}{Nombre de usuario de Github}

\item{repo}{Nombre de repositorio. Si no se especifica utiliza el directorio de trabajo actual.}

\item{state}{Estado del issue. Puede ser 'open' (default), 'closed' o 'all'}

\item{...}{Otros parámetros para el query}
}
\value{
Una lista con clase 'gh_response'
}
\description{
Estas funciones son wrappers de gh::gh para obtener más rápidamente el resultado de una búsqueda
}
\examples{
gh_repos_issues(owner = "calderonsamuel", repo = "casemisc")
}
