% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_row_numbers.R
\name{has_list_columns}
\alias{has_list_columns}
\title{Revisar si un data.frame tiene contiene list.columns}
\usage{
has_list_columns(dat)
}
\arguments{
\item{dat}{Un data.frame}
}
\value{
Un data.frame con la primera columna conteniendo el número de filas
}
\description{
Revisar si un data.frame tiene contiene list.columns
}
